Remove card_routes: all Card operations are on Decks
As implemented a card only has meaning insofar as
it has_many Decks.  While this setup now strikes
me as (possibly) wrong (doesn't a Deck have many
cards?  Or is it the case that a Card is to be
shared across Decks?  The implementation isn't
entirely clear to me)

Whatever the case, it's confusing to find deck
based methods on something called card_routes.
thus we relocate and terminate.